requirements.txt
Django==4.0.3
django-filter==21.1
django-bootstrap-v5==1.0.11
Pillow==9.0.1
